Refactor and some information text
