+++ Code/Current/QikConsole/Commands/CommandFactory.cs
@@ -1,7 +1,5 @@
 using System;
 using System.CommandLine;
-using System.CommandLine.Invocation;
-using System.IO;
 
 namespace CygSoft.Qik.QikConsole
 {
@@ -23,13 +21,12 @@ namespace CygSoft.Qik.QikConsole
         private readonly NLog.ILogger logger;
         private readonly IProjectFile projectFile;
         private readonly IFileFunctions fileFunctions;
-        private readonly IOutputGenerator outputGenerator;
-        public CommandFactory(IProjectFile projectFile, IOutputGenerator outputGenerator, IFileFunctions fileFunctions, NLog.ILogger logger)
+
+        public CommandFactory(IProjectFile projectFile, IFileFunctions fileFunctions, NLog.ILogger logger)
         {
             this.logger = logger ?? throw new ArgumentNullException($"{nameof(logger)} cannot be null.");
             this.projectFile = projectFile ?? throw new ArgumentNullException($"{nameof(projectFile)} cannot be null.");
             this.fileFunctions = fileFunctions ?? throw new ArgumentNullException($"{nameof(fileFunctions)} cannot be null.");
-            this.outputGenerator = outputGenerator ?? throw new ArgumentNullException($"{nameof(outputGenerator)} cannot be null.");
         }
 
         public Command Create(CommandType commandType)
@@ -37,10 +34,10 @@ namespace CygSoft.Qik.QikConsole
             switch (commandType)
             {
                 case CommandType.Generate:
-                    return new SymbolCommand(projectFile, outputGenerator, fileFunctions, logger).Configure();
+                    return new SymbolCommand(projectFile, fileFunctions, logger).Configure();
 
                 case CommandType.SymbolUpdate:
-                    return new GenerateCommand(projectFile, outputGenerator, fileFunctions, logger).Configure();
+                    return new GenerateCommand(projectFile, fileFunctions, logger).Configure();
                 
                 default:
                     throw new NotImplementedException();
