+++ Code/Current/QikConsole/Commands/GenerateCommand.cs
@@ -1,6 +1,10 @@
 using System;
+using System.IO;
+using System.Text;
+using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.Invocation;
+using CygSoft.Qik.Functions;
 
 namespace CygSoft.Qik.QikConsole
 {
@@ -11,14 +15,14 @@ namespace CygSoft.Qik.QikConsole
         private readonly NLog.ILogger logger;
         private readonly IProjectFile projectFile;
         private readonly IFileFunctions fileFunctions;
-        private readonly IOutputGenerator outputGenerator;
 
-        public GenerateCommand(IProjectFile projectFile, IOutputGenerator outputGenerator, IFileFunctions fileFunctions, NLog.ILogger logger)
+        private Dictionary<string, string> fragmentsDictionary = new Dictionary<string, string>();
+
+        public GenerateCommand(IProjectFile projectFile, IFileFunctions fileFunctions, NLog.ILogger logger)
         {
             this.logger = logger ?? throw new ArgumentNullException($"{nameof(logger)} cannot be null.");
             this.projectFile = projectFile ?? throw new ArgumentNullException($"{nameof(projectFile)} cannot be null.");
             this.fileFunctions = fileFunctions ?? throw new ArgumentNullException($"{nameof(fileFunctions)} cannot be null.");
-            this.outputGenerator = outputGenerator ?? throw new ArgumentNullException($"{nameof(outputGenerator)} cannot be null.");
         }
 
         public Command Configure()
@@ -48,7 +52,17 @@ namespace CygSoft.Qik.QikConsole
             {
                 try
                 {
-                    GenerateOutputFiles(filePath);
+                    WriteLine("Generating output files...");
+
+                    fragmentsDictionary = new Dictionary<string, string>();
+                    var project = projectFile.Read(filePath);
+                    
+                    GenerateFragments(filePath, project);
+                    GenerateDocuments(filePath, project);
+
+                    ForegroundColor = ConsoleColor.Green;
+                    WriteLine("...Success!");
+                    ForegroundColor = ConsoleColor.White;
                 }
                 catch (Exception ex)
                 {
@@ -58,13 +72,55 @@ namespace CygSoft.Qik.QikConsole
             }
         }
 
-        private void GenerateOutputFiles(string filePath)
+        public void GenerateFragments(string path, Project project)
         {
-            WriteLine("Generating output files...");
-            outputGenerator.Generate(filePath);
-            ForegroundColor = ConsoleColor.Green;
-            WriteLine("...Success!");
-            ForegroundColor = ConsoleColor.White;
+            var scriptPath = Path.Combine(Path.GetDirectoryName(path), project.ScriptPath);
+            var script = fileFunctions.ReadTextFile(scriptPath);
+            var interpreter = new Interpreter();
+            var symbolTerminal = interpreter.Interpret(new FunctionFactory(), script);
+            var terminal = new PlaceholderTerminal(symbolTerminal, "@{", "}");
+
+            foreach (var input in project.Inputs)
+            {
+                terminal.SetSymbolValue(input.Symbol, input.Value);
+            }
+            
+            foreach (var frag in project.Fragments)
+            {
+                var fullPath = Path.Combine(Path.GetDirectoryName(path), frag.Path);
+                var templateText = fileFunctions.ReadTextFile(fullPath);
+
+                foreach (var placeholder in terminal.Placeholders)
+                {
+                    templateText = templateText.Replace(placeholder, terminal.GetPlaceholderValue(placeholder));
+                }
+
+                fragmentsDictionary.Add(frag.Id, templateText);
+            }
+        }
+
+        public void GenerateDocuments(string path, Project project)
+        {
+            foreach (var document in project.Documents)
+            {
+                StringBuilder builder = new StringBuilder();
+                foreach(var structure in document.Structure)
+                {
+                    if (fragmentsDictionary.ContainsKey(structure))
+                    {
+                        builder.AppendLine(fragmentsDictionary[structure]);
+                    }
+                }
+                
+                foreach (var outputPath in document.OutputFilePaths)
+                {
+                    var filePath = fileFunctions.GetRootedFilePath(path, outputPath);
+
+                    if (fileFunctions.FileExists(filePath)) fileFunctions.DeleteFile(filePath);
+                    
+                    fileFunctions.WriteTextFile(filePath, builder.ToString());
+                }
+            }
         }
 
         private static void DisplayWelcomeHeader()
