--- Code/Current/QikConsole/IOutputGenerator.cs
@@ -1,8 +1,0 @@
-
-namespace CygSoft.Qik.QikConsole
-{
-    public interface IOutputGenerator
-    {
-        void Generate(string path);
-    }
-}
